Make all remaining dropColumn calls conditional

+++ database/migrations/2024_10_08_173537_alter_orden_compra_cab_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('orden_compra_cab', function (Blueprint $table) {
             // Eliminar la relaciÃ³n y la columna proveedor_id
-            $table->dropColumn('proveedor_id');
+            if (Schema::hasColumn('orden_compra_cab', 'proveedor_id')) $table->dropColumn('proveedor_id');
         });
     }
