+++ database/migrations/2025_12_27_205556_remove_user_cierre_id_from_apertura_cierre_caja.php
@@ -10,7 +10,7 @@ return new class extends Migration {
     {
         Schema::table('apertura_cierre_caja', function (Blueprint $table) {
             if (Schema::hasColumn('apertura_cierre_caja', 'user_cierre_id')) {
-                $table->dropColumn('user_cierre_id');
+                if (Schema::hasColumn('apertura_cierre_caja', 'user_cierre_id')) $table->dropColumn('user_cierre_id');
             }
         });
     }
