+++ database/migrations/2026_05_03_113920_add_deposito_id_to_compra_ventas.php
@@ -23,11 +23,11 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('compra_det', function (Blueprint $table) {
-            $table->dropColumn('deposito_id');
+            if (Schema::hasColumn('compra_det', 'deposito_id')) $table->dropColumn('deposito_id');
         });
 
         Schema::table('ventas_det', function (Blueprint $table) {
-            $table->dropColumn('deposito_id');
+            if (Schema::hasColumn('ventas_det', 'deposito_id')) $table->dropColumn('deposito_id');
         });
     }
 };
